Move AbstractColumnReaderTest to io.trino.parquet.reader
